view expense split into two components
